+++ src/wai/annotations/core/__init__.py
@@ -13,4 +13,4 @@ from ._SeparateImageWriter import SeparateImageWriter
 from ._LibrarySettings import LibrarySettings, get_settings, set_settings
 from ._typing import InternalFormat
 from ._Writer import Writer
-from ._ZeroAreaDiscarder import ZeroAreaDiscarder
+from ._DimensionDiscarder import DimensionDiscarder