+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Upgraded wai.common requirement to v0.0.27.
 - Separated main settings from library settings.
+- Changed ``--include-zero-area`` flag to min/max width/height/area flags.
 
 0.2.2 (2020-03-11)
 ------------------